Update the rating schema and entire rating controllers

+++ src/controllers/rating_controller.py
@@ -1,7 +1,9 @@
 from main import db
 from flask import Blueprint, request, jsonify
+from marshmallow.exceptions import ValidationError
 from models.rating import Rating
 from models.movie import Movie
+from models.user import User
 from schemas.rating_schema import rating_schema, ratings_schema
 
 # Initialises flask blueprint for ratings, prefix is nested and registered with users bp
@@ -11,40 +13,47 @@ ratings_bp = Blueprint('ratings', __name__)
 @ratings_bp.route("/", methods=["GET"])
 def get_ratings(user_id):
     '''GET endpoint/handler for fetching specified users movie ratings'''
+
     # Queries rating instance from the DB
     ratings = Rating.query.filter_by(user_id=user_id).all()
-    # Serialises queried rating instances from DB with marshmallow schema into Python DST
 
+    # Query the user using get by user_id from DB
+    user = User.query.get(user_id)
+
+    # Checks if there are movies rated
     if len(ratings) < 1:
-        return jsonify(message="You have not rated a movie yet."), 404
+        return jsonify(message=f"No ratings found for user with ID of {user_id}, please rate a movie"), 404
 
+    # Prepare ratings_schema for response
     response = ratings_schema.dump(ratings)
     # Returns the serialised data into JSON format for response
-    return jsonify(response)
+    return jsonify(message=f"{len(ratings)} movies rated for {user.username}", ratings=response)
 
 
 @ratings_bp.route("/movies/<int:movie_id>/", methods=["POST"])
 def add_movie_rating(user_id, movie_id):
     '''POST endpoint/handler for adding a movie rating of the specified user'''
 
-    rating_body_data = rating_schema.load(request.json)
+    # Validating rating request body data with schema
+    try:
+        # If successful, load the request body data
+        rating_body_data = rating_schema.load(request.json)
+    except ValidationError as error:
+        # If fail, return error message
+        return jsonify(error.messages), 400
 
-    # Get the movie from DB
+    # Query movie from DB based on movie id
     movie = Movie.query.get(movie_id)
-
+    # Checks if the movie exists
     if not movie:
-        return jsonify(message="Movie not found"), 404
+        return jsonify(message=f"Movie with ID {movie_id} cannot be found, please try again"), 404
 
+    # Query an existing rating for a movie from DB filtered by both user_id and movie_id
     existing_rating = Rating.query.filter_by(
         user_id=user_id, movie_id=movie_id).first()
-
+    # Checks if a rating for a mvovie already exists to avoid duplication
     if existing_rating:
-        return jsonify(message="Movie already rated"), 409
-    
-    rating_score = rating_body_data.get("rating_score")
-
-    if not 1 <= rating_score <= 5:
-        return jsonify(message="Invalid rating. Rating should be between 1 and 5."), 400
+        return jsonify(message=f"{movie.title} has already been rated"), 409
 
     # Create new rating instance
     new_rating = Rating(
@@ -59,45 +68,68 @@ def add_movie_rating(user_id, movie_id):
 
     # Return the new rating as a JSON response
     response = rating_schema.dump(new_rating)
-    return jsonify(response), 201
+    return jsonify(message=f"{movie.title} added to ratings!", rating=response), 201
 
 
 @ratings_bp.route("/movies/<int:movie_id>/", methods=["PUT"])
 def update_movie_rating(user_id, movie_id):
     '''PUT endpoint/handler for updating a movie's rating of the specified user'''
-    rating_body_data = rating_schema.load(request.json)
 
+    # Validating rating request body data with schema
+    try:
+        # If successful, load the request body data
+        rating_body_data = rating_schema.load(request.json)
+    except ValidationError as error:
+        # If fail, return error message
+        return jsonify(error.messages), 400
+    
+    # Query movie from DB based on movie id
+    movie = Movie.query.get(movie_id)
+    # Checks if the movie exists
+    if not movie:
+        return jsonify(message=f"Movie with ID {movie_id} cannot be found, please try again"), 404
+
+    # Query an existing rating for a movie from DB filtered by both user_id and movie_id
     existing_rating = Rating.query.filter_by(
         user_id=user_id, movie_id=movie_id).first()
-
+    # Checks if a rating for a movie already exists for the rating to be updated
     if not existing_rating:
-        return jsonify(message="No existing rating found for this movie by this user."), 404
-    
-    rating_score = rating_body_data.get("rating_score")
+        return jsonify(message=f"No existing rating found for {movie.title} by user with ID of {user_id}"), 404
 
-    if not 1 <= rating_score <= 5:
-        return jsonify(message="Invalid rating. Rating should be between 1 and 5."), 400
-
+    # Update the existing rating with the new rating
     existing_rating.rating_score = rating_body_data["rating_score"]
 
+    # Commit updated changes to DB
     db.session.commit()
 
+    # Return the updated rating as a JSON response
     response = rating_schema.dump(existing_rating)
-    return jsonify(response), 200
+    return jsonify(message=f" Rating for {movie.title} successfully updated!", rating=response), 200
 
 
 @ratings_bp.route("/movies/<int:movie_id>/", methods=["DELETE"])
 def remove_movie_rating(user_id, movie_id):
     '''DELETE endpoint/handler for removing a movie rating of the specified user'''
+
+    # Query movie from DB based on movie id
+    movie = Movie.query.get(movie_id)
+    # Checks if the movie exists
+    if not movie:
+        return jsonify(message=f"Movie with ID {movie_id} cannot be found, please try again"), 404
+    
+    # Query an existing rating for a movie from DB filtered by both user_id and movie_id
     existing_rating = Rating.query.filter_by(
         user_id=user_id, movie_id=movie_id).first()
-
+    # Checks if a rating for a movie already exists for the rating to be removed
     if not existing_rating:
-        return jsonify(message="No existing rating found for this movie by this user."), 404
+        return jsonify(message=f"No existing rating found for {movie.title} by user with ID of {user_id}"), 404
 
+    # Save rating before deletion for response
     response = rating_schema.dump(existing_rating)
 
+    # Delete exisiting rating and commit changes to DB
     db.session.delete(existing_rating)
     db.session.commit()
 
-    return jsonify(message="Movie rating sucessfully removed.", deleted_rating=response), 200
+    # Return message and deleted_rating response as JSON
+    return jsonify(message=f"{movie.title} sucessfully removed!", deleted_rating=response), 200
